Backend/requirements.txt
asgiref==3.7.2
Django==4.1
django-cors-headers==4.2.0
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
environ==1.0
psycopg2-binary==2.9.7
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
boto3==1.26.110
botocore==1.29.110
django-storages==1.13.2
